+++ src/js/lotto/index.js
@@ -1,0 +1,4 @@
+import validateLotto from './validateLotto.js';
+import makeLotto from './makeLotto.js';
+
+export { validateLotto, makeLotto };
