+++ src/js/lotto/validateLotto.js
@@ -1,17 +1,6 @@
-const LOTTO_MAX_NUMBER = 45;
-const LOTTO_MIN_NUMBER = 1;
-const LOTTO_LENGTH = 6;
+import { LOTTO_LENGTH, LOTTO_MAX_NUMBER, LOTTO_MIN_NUMBER } from '../constants/lotto.js';
 
-function makeLotto() {
-  const balls = new Array(LOTTO_MAX_NUMBER).fill(null).map((value, index) => index + 1);
-  const lotto = new Array(LOTTO_LENGTH).fill(null).map(() => {
-    const ballIndex = Math.floor(Math.random() * (balls.length - 1));
-    return balls.splice(ballIndex, 1)[0];
-  });
-  return lotto;
-}
-
-function validateLotto(numbers) {
+export default function validateLotto(numbers) {
   if (numbers.length !== LOTTO_LENGTH) throw new Error(`Lotto should have ${LOTTO_LENGTH} numbers`);
 
   const isOutOfRangeNumber =
@@ -22,5 +11,3 @@ function validateLotto(numbers) {
   const hasDuplicateNumber = new Set(numbers).size !== LOTTO_LENGTH;
   if (hasDuplicateNumber) throw new Error('Lotto should not have duplicate numbers');
 }
-
-export { makeLotto, validateLotto };
